space_stations/pyproject.toml
[tool.poetry]
name = "space-stations"
version = "0.1.0"
description = ""
authors = ["Стацюк Виталий"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
sqlparse = "^0.4.3"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
pytz = "^2022.6"
drf-spectacular = "^0.24.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
pytest-pythonpath = "0.7.3"
mixer = "^7.2.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
